Mockup for better error handler
Currently the user must get and parse the response text from Telegram, the idea is that the package will do that job from now on.
Note1: This will introduce BC, so don't merge until v3 is ready to be launched.
Note2: I'm not quite happy with the result yet, haven't thought it 100% out.
Note3: A lot of tests will be failing now, fix those as well